Created submodules that import/set versions of specs, platforms based on detected pydantic version. Other model ops are performed based on pydantic version
